Refactor code of GridList
