+++ client/src/app/styles/index.scss
@@ -69,6 +69,4 @@
 @import '../../common/components/Forms/ErrorMessage.scss';
 @import '../../common/components/SwitchButton.scss';
 @import '../../modules/dashboard/Dashboard.scss';
-@import '../../common/components/Elements.scss';
 @import '../../common/components/ListModeItem.scss';
-@import '../../common/components/ListMode.scss';
