--- client/src/common/components/Elements.scss
@@ -1,31 +1,0 @@
-.elements {
-  width: 100%;
-  margin: 16px 0 25px 0;
-  border: 0;
-
-  &__heading {
-    padding: 8px 0;
-    display: flex;
-    align-items: center;
-    font-size: 16px;
-    line-height: 1;
-    color: $text-normal;
-    border-bottom: 1px solid $border-light;
-    text-transform: uppercase;
-    font-weight: 700;
-    letter-spacing: 0.2px;
-
-    svg {
-      height: 14px;
-      margin-right: 8px;
-
-      path {
-        fill: $text-normal;
-      }
-    }
-  }
-
-  &__body {
-    padding: 16px 0 0;
-  }
-}
