+++ client/src/common/components/GridList.scss
@@ -1,21 +1,66 @@
-.tiles-mode {
-  @include breakpoint($small) {
-    display: grid;
-    grid-template-columns: repeat(auto-fill, 180px);
-    grid-gap: 30px 30px;
+.elements {
+  width: 100%;
+  margin: 16px 0 25px 0;
+  border: 0;
+
+  &__heading {
+    padding: 8px 0;
+    display: flex;
+    align-items: center;
+    font-size: 16px;
+    line-height: 1;
+    color: $text-normal;
+    border-bottom: 1px solid $border-light;
+    text-transform: uppercase;
+    font-weight: 700;
+    letter-spacing: 0.2px;
+
+    svg {
+      height: 14px;
+      margin-right: 8px;
+
+      path {
+        fill: $text-normal;
+      }
+    }
   }
 
-  &__item {
-    margin: 0 auto 20px;
-    width: 100%;
-    height: 96px;
+  &__body {
+    padding: 16px 0 0;
+  }
 
-    @include breakpoint($small) {
-      margin: 0 auto 0;
-    }
+  &__element {
+    width: 100%;
+    margin: 0 auto 20px;
 
     a {
       text-decoration: none;
+    }
+  }
+
+  &__list {
+    .elements__element {
+      height: 130px;
+
+      @include breakpoint($small) {
+        height: 80px;
+      }
+    }
+  }
+
+  &__tiles {
+    @include breakpoint($small) {
+      display: grid;
+      grid-template-columns: repeat(auto-fill, 180px);
+      grid-gap: 30px 30px;
+    }
+
+    .elements__element {
+      height: 96px;
+
+      @include breakpoint($small) {
+        margin: 0 auto 0;
+      }
     }
   }
