--- client/src/common/components/ListMode.scss
@@ -1,27 +1,0 @@
-.list-mode {
-  &__item {
-    width: 100%;
-    height: 130px;
-    margin: 0 auto 20px;
-
-    a {
-      text-decoration: none;
-    }
-
-    @include breakpoint($small) {
-      height: 80px;
-    }
-  }
-
-  &__button {
-    background: transparent;
-    border: 0;
-    cursor: pointer;
-    display: block;
-    height: 100%;
-    margin: 0;
-    outline: 0;
-    padding: 0;
-    width: 100%;
-  }
-}
